backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for the web application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==5.0",
    "djangorestframework==3.16.0",
    "django-allauth>=0.65.0",
    "dj-rest-auth>=3.2.2",
    "requests-oauthlib>=1.3.1",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "django-cors-headers==4.2.0",
    "django-environ>=0.12.0",
    "pyjwt>=2.10.1",
    "cryptography>=46.0.1",
    "whitenoise>=6.11.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
]
